Rebase onto master, clean up some old invocations of the swarm_train and moving the dict stuff into utils as opposed to core (for simplification)
