+++ tests/test_util.py
@@ -29,3 +29,12 @@ def test_transpose_generator_in(size):
 @pytest.mark.parametrize("size", (3, 5, 10))
 def test_transpose_reverse(size):
     assert util.transpose(data_out(size)) == list(data_in(size))
+
+
+def test_merges():
+    d1 = {"a": 3}
+    d2 = {"b": 4}
+    d3 = {"c": 5, "d": 6}
+    assert util.merge_dicts(d1, d2, d3) == {
+        "a": 3, "b": 4, "c": 5, "d": 6
+    }
